use the builtins factory in the C frontend
