+++ src/ansi-c/builtin_factory.cpp
@@ -12,13 +12,11 @@ Author: Daniel Kroening, kroening@kroening.com
 #include "ansi_c_parser.h"
 #include "ansi_c_typecheck.h"
 
+#include <util/config.h>
+
 #include <cstring>
 #include <ostream>
 #include <sstream>
-
-const char windows_headers[]=
-  "int __noop();\n"
-  "int __assume(int);\n";
 
 //! Advance to the next line
 static const char *next_line(const char *line)
@@ -93,7 +91,8 @@ static bool convert(
 
   symbol_tablet new_symbol_table;
 
-  // very recursive!
+  // this is recursive -- builtin_factory is called
+  // from the typechecker
   if(ansi_c_typecheck(
     ansi_c_parser.parse_tree,
     new_symbol_table,
@@ -137,33 +136,96 @@ bool builtin_factory(
   std::string code;
   ansi_c_internal_additions(code);
   s << code;
-  
-  if(find_pattern(pattern, gcc_builtin_headers_generic, s))
-    return convert(identifier, s, symbol_table, mh);
 
-  if(find_pattern(pattern, gcc_builtin_headers_mem_string, s))
-    return convert(identifier, s, symbol_table, mh);
+  // this is Visual C/C++ only
+  if(config.ansi_c.os==configt::ansi_ct::ost::OS_WIN)
+  {
+    if(find_pattern(pattern, windows_builtin_headers, s))
+      return convert(identifier, s, symbol_table, mh);
+  }
 
-  if(find_pattern(pattern, gcc_builtin_headers_omp, s))
-    return convert(identifier, s, symbol_table, mh);
+  // ARM stuff
+  if(config.ansi_c.mode==configt::ansi_ct::flavourt::ARM)
+  {
+    if(find_pattern(pattern, arm_builtin_headers, s))
+      return convert(identifier, s, symbol_table, mh);
+  }
 
-  if(find_pattern(pattern, gcc_builtin_headers_tm, s))
-    return convert(identifier, s, symbol_table, mh);
+  // CW stuff
+  if(config.ansi_c.mode==configt::ansi_ct::flavourt::CODEWARRIOR)
+  {
+    if(find_pattern(pattern, cw_builtin_headers, s))
+      return convert(identifier, s, symbol_table, mh);
+  }
 
-  if(find_pattern(pattern, gcc_builtin_headers_ubsan, s))
-    return convert(identifier, s, symbol_table, mh);
+  // GCC junk stuff, also for CLANG and ARM
+  if(config.ansi_c.mode==configt::ansi_ct::flavourt::GCC ||
+     config.ansi_c.mode==configt::ansi_ct::flavourt::APPLE ||
+     config.ansi_c.mode==configt::ansi_ct::flavourt::ARM)
+  {
+    if(find_pattern(pattern, gcc_builtin_headers_generic, s))
+      return convert(identifier, s, symbol_table, mh);
 
-  if(find_pattern(pattern, gcc_builtin_headers_ia32, s))
-    return convert(identifier, s, symbol_table, mh);
+    if(find_pattern(pattern, gcc_builtin_headers_math, s))
+      return convert(identifier, s, symbol_table, mh);
 
-  if(find_pattern(pattern, gcc_builtin_headers_ia32_2, s))
-    return convert(identifier, s, symbol_table, mh);
+    if(find_pattern(pattern, gcc_builtin_headers_mem_string, s))
+      return convert(identifier, s, symbol_table, mh);
 
-  if(find_pattern(pattern, gcc_builtin_headers_ia32_3, s))
-    return convert(identifier, s, symbol_table, mh);
+    if(find_pattern(pattern, gcc_builtin_headers_omp, s))
+      return convert(identifier, s, symbol_table, mh);
 
-  if(find_pattern(pattern, gcc_builtin_headers_ia32_4, s))
-    return convert(identifier, s, symbol_table, mh);
+    if(find_pattern(pattern, gcc_builtin_headers_tm, s))
+      return convert(identifier, s, symbol_table, mh);
+
+    if(find_pattern(pattern, gcc_builtin_headers_ubsan, s))
+      return convert(identifier, s, symbol_table, mh);
+
+    if(find_pattern(pattern, clang_builtin_headers, s))
+      return convert(identifier, s, symbol_table, mh);
+
+    if(config.ansi_c.arch=="i386" ||
+       config.ansi_c.arch=="x86_64" ||
+       config.ansi_c.arch=="x32")
+    {
+      if(find_pattern(pattern, gcc_builtin_headers_ia32, s))
+        return convert(identifier, s, symbol_table, mh);
+
+      if(find_pattern(pattern, gcc_builtin_headers_ia32_2, s))
+        return convert(identifier, s, symbol_table, mh);
+
+      if(find_pattern(pattern, gcc_builtin_headers_ia32_3, s))
+        return convert(identifier, s, symbol_table, mh);
+
+      if(find_pattern(pattern, gcc_builtin_headers_ia32_4, s))
+        return convert(identifier, s, symbol_table, mh);
+    }
+    else if(config.ansi_c.arch=="arm64" ||
+            config.ansi_c.arch=="armel" ||
+            config.ansi_c.arch=="armhf" ||
+            config.ansi_c.arch=="arm")
+    {
+      if(find_pattern(pattern, gcc_builtin_headers_arm, s))
+        return convert(identifier, s, symbol_table, mh);
+    }
+    else if(config.ansi_c.arch=="mips64el" ||
+            config.ansi_c.arch=="mipsn32el" ||
+            config.ansi_c.arch=="mipsel" ||
+            config.ansi_c.arch=="mips64" ||
+            config.ansi_c.arch=="mipsn32" ||
+            config.ansi_c.arch=="mips")
+    {
+      if(find_pattern(pattern, gcc_builtin_headers_mips, s))
+        return convert(identifier, s, symbol_table, mh);
+    }
+    else if(config.ansi_c.arch=="powerpc" ||
+            config.ansi_c.arch=="ppc64" ||
+            config.ansi_c.arch=="ppc64le")
+    {
+      if(find_pattern(pattern, gcc_builtin_headers_power, s))
+        return convert(identifier, s, symbol_table, mh);
+    }
+  }
 
   return true;
 }
