+++ src/ansi-c/clang_builtin_headers.h
@@ -1,5 +1,3 @@
-typedef float  __gcc_v4sf  __attribute__ ((__vector_size__ (16)));
-
 __gcc_v4sf __builtin_shufflevector(__gcc_v4sf, __gcc_v4sf, ...);
 
 int __builtin_flt_rounds(void);
